requirements.txt
hachoir==3.2.0
Pillow==10.0.1
Telethon==1.30.3